+++ v8/plantuml/CHANGES.md
@@ -1,5 +1,6 @@
 # 1.0.0
-* Renamed `PLANTUML_ARGS` config option to `PLANTUML_FILE_OPTIONS`.
+* Add support for PlantUML PicoWeb server.
+* Rename `PLANTUML_ARGS` config option to `PLANTUML_FILE_OPTIONS`.
 
 # 0.2.0
 * Add `PlantUmlTask.plantuml_manager` so the `plantuml_markdown` plugin can use it.
@@ -8,4 +9,4 @@
 * Update PLANTUML_FILES in conf.py.sample to match the default behaviour.
 
 # 0.1
-* First release
+* First release.